src/api/people + controllers
